Refactoring - split (WIP)
